# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    API for messaging app WASAText
  version: 0.0.1
security:
  - bearerAuth: [ ]
...
paths:
...
  /chats:
    get:
      tags:
        - chats
      operationId: getMyConversations
      summary: Get user chats
      description: "Get all the user's conversations"
      responses:
        "200":
          description: Chats found
          content:
            application/json:
              schema:
                description: chats
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
                minItems: 1
                maxItems: 999
        "204":
          description: "No Chats"
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - chats
      summary: Create new Chat
      description: |
        Send a new message in a conversation. If it's the first one,
        a new conversation will be created
      operationId: createChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirstMessageRequest"
        required: true
      responses:
        '201':
          description: New chat created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /chats/{chatId}/members:
    parameters:
      - $ref: "#/components/parameters/ChatID"
    post:
      tags:
        - chats
      operationId: addToGroup
      summary: Add user to group
      description: "Add an user to a a group. If user is already present, nothing will change"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsernameRequest'
      responses:
        "201":
          description: User added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "200":
          description: User already present
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User is already a member"
    delete:
      tags:
        - chats
      summary: Leave group
      description: "Remove an user from a group. Users can only remove themselves."
      operationId: leaveGroup
      responses:
        "204":
          description: Left Group
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /users/{userId}/chats/{chatId}/messages:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/ChatID"
    post:
      tags:
        - messages
      summary: Send a message
      description: |
        Send a new message in a conversation. If it's the first one,
        a new conversation will be created
      operationId: sendMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
        required: true
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  schemas:
    User:
      title: User
      description: Unique user
      type: object
      properties:
        id:
          description: Unique identifier for the user
          type: integer
          example: 223
          readOnly: True
        username:
          description: username
          type: string
          example: 'mario'
          pattern: "^[a-zA-Z0-9_-]+$"
          minLength: 3
          maxLength: 16
    Chat:
      title: Chat
      description: Generic chat
      type: object
      properties:
        id:
          description: Unique identifier for the chat
          type: integer
          example: 223
          readOnly: True
        isGroup:
          description: identifies chat as a group or private chat
          type: boolean
          example: true
        chatName:
          description: Chat title, user name or group name
          type: string
          example: "Study group"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 3
          maxLength: 24
        lastMsgText:
          description: text of the last message in the chat
          type: string
          example: "The meeting is canceled"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 1
          maxLength: 99
        lastMsgTime:
          description: Date and time when the message is sent
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          example: "2024-10-11T12:14:23Z"
          readOnly: true
          minLength: 20
          maxLength: 20
        lastMsgType:
          description: type of the message, either text or image
          type: string
          enum: ["text", "image"]
          example: "text"
    Message:
      title: Message
      description: Single message
      type: object
      properties:
        id:
          description: Unique identifier for the message
          type: integer
          example: 223
          readOnly: True
        type:
          description: type of the message, either text or image
          type: string
          enum: [ "text", "image" ]
          example: "text"
        text:
          description: message text
          type: string
          example: "hello world"
          pattern: "^[a-zA-Z0-9_ -]+$"
          minLength: 1
          maxLength: 250
        chatId:
          description: Identifier for the chat
          type: integer
          example: 223
        senderId:
          description: Identifier for the sender
          type: integer
          example: 223
        isForward:
          type: boolean
          description: Whether this message is forwarded from another chat.
          example: false
        replyTo:
          description: Identifier for the message this one is replying to
          type: integer
          example: 223
        createdAt:
          description: Date and time when the message is sent
          type: string
          format: date-time
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          example: "2024-10-11T12:14:23Z"
          readOnly: true
          minLength: 20
          maxLength: 20
...
    UsernameRequest:
      title: Username Request
      description: Request consists in a string to change the username
      type: object
      properties:
        username:
          description: username
          type: string
          example: 'mario'
          pattern: "^[a-zA-Z0-9_-]+$"
          minLength: 3
          maxLength: 16
    FirstMessageRequest:
      type: object
      properties:
        text:
          type: string
          description: The text content of the message.
          example: "Hello, team!"
        type:
          type: string
          description: The type of message (e.g., text, image, video).
          example: "text"
        mediaUrl:
          type: string
          format: uri
          description: URL to any attached media.
          example: "https://cdn.example.com/media/abc123.jpg"
        chatName:
          type: string
          description: Name of the chat to create if new.
          example: "Project X"
        receivers:
          type: array
          description: List of user IDs to receive the message.
          items:
            type: integer
          example: [ 123, 456, 789 ]
        isForward:
          type: boolean
          description: Whether this message is forwarded from another chat.
          example: false
      required:
        - text
        - type
        - receivers
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "The request could not be processed"
        code:
          type: integer
          example: 400
  parameters:
    UserID:
      schema:
        type: number
        format: integer
        example: 134
      name: userId
      in: path
      required: true
      description: User identifier
    ChatID:
      schema:
        type: number
        format: integer
        example: 134
      name: chatId
      in: path
      required: true
      description: Chat identifier
...
  responses:
...
    UnauthorizedError:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer